requirements.txt
# --- FastAPI Framework ---
fastapi==0.110.0
uvicorn[standard]==0.30.0

# --- Qdrant Vector DB Client ---
qdrant-client==1.11.1

# --- HTTP Requests ---
httpx==0.25.2    # ✅ Common compatible version
requests==2.32.3

# --- Data Models & Validation ---
#pydantic==1.10.15
pydantic==2.7.4         # ✅ Compatible with FastAPI 0.110.0
pydantic-core==2.18.4   # ✅ Needed for Pydantic v2
python-dotenv==1.0.1

# --- PDF & DOCX Parsing ---
PyMuPDF==1.24.7
python-docx==1.1.2

# --- File Upload & Multipart Handling ---
python-multipart==0.0.9

# --- Async Utilities ---
aiofiles==24.1.0

# --- Logging & Monitoring ---
loguru==0.7.2

# --- Optional: Development Utilities ---
black==24.4.2
isort==5.13.2
flake8==7.1.0

# --- Testing ---
pytest==8.2.2

# --- Ollama Integration ---
ollama==0.1.7

# --- Environment Management ---
python-decouple==3.8
